moved ExtractVolumeFromPath
